update progress (#3)
* seprate img

* bme280-oled

* last_str

* progress

+++ init.lua
@@ -1,3 +1,7 @@
+if file.exists("cmd.txt") then
+  return
+end
+
 if file.exists("update.txt") then
   -- enter update mode
   file.remove("update.txt")
